update TwitchAPIPull to pull and map the data in a single location
getPOJO now gets a link and a POJO Class that it uses to make the
correct POJO. The return type is Object, so a cast to the correct POJO
must be done. 

also move data fields form Channel to ChannelsFollowsUser
May decide to remove ChannelsFollowsUser and use a Channel Instead.